requirements.txt
Pillow
numpy
huggingface-hub
matplotlib
onnxruntime
opencv-python